webadmin: New disk dialog does not validate the size value
The size field is defined as as an entity model of type long and so the
integer validator does not validate it at all making it possible for numbers
less than the minimum (including negative number and zero) as well as
numbers higher than the maximum to appear as validated, this was fixed
by replacing the entity model for size from type long to integer

Bug-Url: https://bugzilla.redhat.com/1115471

+++ frontend/webadmin/modules/gwt-common/src/main/java/org/ovirt/engine/ui/common/widget/uicommon/storage/DisksAllocationItemView.java
@@ -160,7 +160,7 @@ public class DisksAllocationItemView extends Composite implements HasEditorDrive
         driver.edit(object);
 
         diskAliasLabel.asValueBox().setValue(object.getAlias().getEntity());
-        diskSizeLabel.asValueBox().setValue((new DiskSizeRenderer<Long>(SizeConverter.SizeUnit.GB).render(
+        diskSizeLabel.asValueBox().setValue((new DiskSizeRenderer<Integer>(SizeConverter.SizeUnit.GB).render(
                 object.getSize().getEntity())));
 
         object.getVolumeType().setSelectedItem(((DiskImage) object.getDisk()).getVolumeType());
